react-search-agent/pyproject.toml
[project]
name = "react-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.7",
    "langchain-tavily>=0.2.11",
    "openai>=1.108.0",
    "python-dotenv>=1.1.1",
]
